+++ src/main/java/org/riekr/jloga/cobol/CobolTextSource.java
@@ -11,9 +11,6 @@ import java.io.IOException;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
-import java.util.function.BooleanSupplier;
 import java.util.function.Supplier;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
@@ -21,12 +18,8 @@ import java.util.stream.Stream;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.riekr.jloga.io.FileProgressInputStream;
-import org.riekr.jloga.io.FilteredTextSource;
-import org.riekr.jloga.io.IndexingException;
 import org.riekr.jloga.io.ProgressListener;
-import org.riekr.jloga.io.VolatileTextSource;
-import org.riekr.jloga.pmem.PagedJavaList;
-import org.riekr.jloga.search.SearchPredicate;
+import org.riekr.jloga.io.ReloadableVolatileTextSource;
 
 import net.sf.JRecord.Common.FieldDetail;
 import net.sf.JRecord.Common.IFieldDetail;
@@ -43,7 +36,7 @@ import net.sf.JRecord.Numeric.ICopybookDialects;
 import net.sf.JRecord.def.IO.builders.ICobolIOBuilder;
 import net.sf.cb2xml.def.Cb2xmlConstants;
 
-public class CobolTextSource extends VolatileTextSource {
+public class CobolTextSource extends ReloadableVolatileTextSource {
 
 	private static final String _DELIM = "|";
 
@@ -55,8 +48,6 @@ public class CobolTextSource extends VolatileTextSource {
 	private final @NotNull Integer _splitCopybook;
 	private final @NotNull Integer _dialect;
 
-	private Future<?> _indexing;
-
 	public CobolTextSource(
 			@NotNull String copybook,
 			@NotNull String datafile,
@@ -67,7 +58,6 @@ public class CobolTextSource extends VolatileTextSource {
 			@Nullable Integer dialect,
 			@Nullable Supplier<ProgressListener> progressListenerSupplier
 	) {
-		super(PagedJavaList.ofStrings());
 		_copybook = copybook;
 		_datafile = datafile;
 		_copybookFileFormat = copybookFileFormat == null ? Cb2xmlConstants.USE_STANDARD_COLUMNS : copybookFileFormat;
@@ -75,96 +65,56 @@ public class CobolTextSource extends VolatileTextSource {
 		_fileOrganization = fileOrganization == null ? IFileStructureConstants.IO_FIXED_LENGTH : fileOrganization;
 		_splitCopybook = splitCopybook == null ? CopybookLoader.SPLIT_NONE : splitCopybook;
 		_dialect = dialect == null ? ICopybookDialects.FMT_MAINFRAME : dialect;
-		_indexing = defaultAsyncIO(() -> reload(progressListenerSupplier == null ? () -> ProgressListener.NOP : progressListenerSupplier));
+		doFirstLoad(progressListenerSupplier);
 	}
 
 	@Override
-	public boolean supportsReload() {
-		return true;
-	}
+	protected void doReload(Supplier<ProgressListener> progressListenerSupplier) throws IOException {
+		final ICobolIOBuilder builder = JRecordInterface1.COBOL.newIOBuilder(_copybook)
+				.setFileOrganization(_fileOrganization)
+				.setSplitCopybook(_splitCopybook)
+				.setDialect(_dialect)
+				.setCopybookFileFormat(_copybookFileFormat)
+				.setFont(_font);
 
-	public Future<?> requestReload(Supplier<ProgressListener> progressListenerSupplier) {
-		if (_indexing == null || _indexing.isDone())
-			_indexing = defaultAsyncIO(() -> reload(progressListenerSupplier));
-		return _indexing;
-	}
+		// final Map<Integer, RecordDetail> records = new LinkedHashMap<>();
+		final Map<String, IFieldDetail> header = new LinkedHashMap<>();
+		final LayoutDetail layout = builder.getLayout();
+		for (final RecordDetail recordDetail : layout.getRecordsAsList()) {
+			System.out.println("Record: " + recordDetail.getRecordName() + '/' + recordDetail.getParentRecordIndex());
+			// records.put(recordDetail.getSourceIndex(), recordDetail);
+			for (final FieldDetail field : recordDetail.getFields()) {
+				if (header.put(field.getName(), field) != null)
+					System.err.println("Duplicate field: " + field.getName());
+			}
+		}
+		RecordDecider decider = layout.getDecider();
+		if (decider == null)
+			add(String.join(_DELIM, header.keySet()));
+		else
+			add("R" + _DELIM + String.join(_DELIM, header.keySet()));
 
-	@Override
-	public void close() {
-		_indexing.cancel(true);
-	}
+		AbstractLineReader reader = null;
+		try (FileProgressInputStream fis = new FileProgressInputStream(progressListenerSupplier.get(), _datafile)) {
+			reader = builder.newReader(fis);
 
-	@Override
-	public boolean isIndexing() {
-		return !_indexing.isDone();
-	}
-
-	@Override
-	public void search(SearchPredicate predicate, FilteredTextSource out, ProgressListener progressListener, BooleanSupplier running) throws ExecutionException, InterruptedException {
-		_indexing.get();
-		super.search(predicate, out, progressListener, running);
-	}
-
-	@Override
-	public int getLineCount() throws ExecutionException, InterruptedException {
-		_indexing.get();
-		return super.getLineCount();
-	}
-
-	@Override
-	public void reload(Supplier<ProgressListener> progressListenerSupplier) {
-		try {
-			clear();
-
-			final ICobolIOBuilder builder = JRecordInterface1.COBOL.newIOBuilder(_copybook)
-					.setFileOrganization(_fileOrganization)
-					.setSplitCopybook(_splitCopybook)
-					.setDialect(_dialect)
-					.setCopybookFileFormat(_copybookFileFormat)
-					.setFont(_font);
-
-			// final Map<Integer, RecordDetail> records = new LinkedHashMap<>();
-			final Map<String, IFieldDetail> header = new LinkedHashMap<>();
-			final LayoutDetail layout = builder.getLayout();
-			for (final RecordDetail recordDetail : layout.getRecordsAsList()) {
-				System.out.println("Record: " + recordDetail.getRecordName() + '/' + recordDetail.getParentRecordIndex());
-				// records.put(recordDetail.getSourceIndex(), recordDetail);
-				for (final FieldDetail field : recordDetail.getFields()) {
-					if (header.put(field.getName(), field) != null)
-						System.err.println("Duplicate field: " + field.getName());
+			AbstractLine line;
+			while ((line = reader.read()) != null) {
+				final Stream<String> values = header.values().stream()
+						.map(line::getFieldValue)
+						.map(IFieldValue::asString);
+				if (decider == null)
+					add(values.collect(joining(_DELIM)));
+				else {
+					add(Stream.concat(
+							IntStream.of(decider.getPreferedIndex(line)).mapToObj(Integer::toString),
+							values
+					).collect(joining(_DELIM)));
 				}
 			}
-			RecordDecider decider = layout.getDecider();
-			if (decider == null)
-				add(String.join(_DELIM, header.keySet()));
-			else
-				add("R" + _DELIM + String.join(_DELIM, header.keySet()));
-
-			AbstractLineReader reader = null;
-			try (FileProgressInputStream fis = new FileProgressInputStream(progressListenerSupplier.get(), _datafile)) {
-				reader = builder.newReader(fis);
-
-				AbstractLine line;
-				while ((line = reader.read()) != null) {
-					final Stream<String> values = header.values().stream()
-							.map(line::getFieldValue)
-							.map(IFieldValue::asString);
-					if (decider == null)
-						add(values.collect(joining(_DELIM)));
-					else {
-						add(Stream.concat(
-								IntStream.of(decider.getPreferedIndex(line)).mapToObj(Integer::toString),
-								values
-						).collect(joining(_DELIM)));
-					}
-				}
-			} finally {
-				if (reader != null)
-					reader.close();
-			}
-		} catch (IOException e) {
-			e.printStackTrace(System.err);
-			throw new IndexingException("Error reloading " + _datafile, e);
+		} finally {
+			if (reader != null)
+				reader.close();
 		}
 	}