+++ src/main/java/org/riekr/jloga/io/VolatileTextSource.java
@@ -3,7 +3,6 @@ package org.riekr.jloga.io;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReadWriteLock;
@@ -15,11 +14,14 @@ import org.riekr.jloga.react.IntBehaviourSubject;
 import org.riekr.jloga.react.Observer;
 import org.riekr.jloga.react.Unsubscribable;
 
+/**
+ * Like {@link TempFilteredTextSource} but faster and only for root panes.
+ */
 public class VolatileTextSource implements TextSource {
 
-	private final IntBehaviourSubject _lineCountSubject = new IntBehaviourSubject();
-	private final List<String>        _data;
-	private final ReadWriteLock       _readWriteLock    = new ReentrantReadWriteLock();
+	protected final IntBehaviourSubject _lineCountSubject = new IntBehaviourSubject();
+	private final   List<String>        _data;
+	private final   ReadWriteLock       _readWriteLock    = new ReentrantReadWriteLock();
 
 	public VolatileTextSource(@NotNull List<String> backend) {
 		_data = backend;
@@ -42,7 +44,7 @@ public class VolatileTextSource implements TextSource {
 		_lineCountSubject.next(0);
 	}
 
-	public final void add(String line) {
+	public void add(String line) {
 		final Lock lock = _readWriteLock.writeLock();
 		lock.lock();
 		try {
@@ -54,7 +56,7 @@ public class VolatileTextSource implements TextSource {
 	}
 
 	@Override
-	public String getText(int line) {
+	public final String getText(int line) {
 		final Lock lock = _readWriteLock.readLock();
 		lock.lock();
 		try {
@@ -67,17 +69,38 @@ public class VolatileTextSource implements TextSource {
 	}
 
 	@Override
-	public Future<Integer> requestIntermediateLineCount(IntConsumer consumer) {
+	public final String[] getText(int fromLine, int count) {
+		int toLinePlus1 = Math.min(fromLine + count, Math.toIntExact(_data.size()));
+		final String[] lines = new String[toLinePlus1 - fromLine + 1];
+		final Lock lock = _readWriteLock.readLock();
+		lock.lock();
+		try {
+			for (int i = fromLine; i <= toLinePlus1; i++) {
+				String line;
+				if (i < 0 || i >= _data.size())
+					line = "";
+				else
+					line = _data.get(i);
+				lines[i - fromLine] = line;
+			}
+		} finally {
+			lock.unlock();
+		}
+		return lines;
+	}
+
+	@Override
+	public final Future<Integer> requestIntermediateLineCount(IntConsumer consumer) {
 		return _lineCountSubject.once(Observer.async(consumer::accept));
 	}
 
 	@Override
-	public Unsubscribable subscribeLineCount(IntConsumer consumer) {
+	public final Unsubscribable subscribeLineCount(IntConsumer consumer) {
 		return _lineCountSubject.subscribe(consumer::accept);
 	}
 
 	@Override
-	public int getLineCount() throws ExecutionException, InterruptedException {
+	public int getLineCount() {
 		return _data.size();
 	}
 }