+++ src/test/java/org/riekr/jloga/transform/HeaderDetectorTest.java
@@ -3,21 +3,21 @@ package org.riekr.jloga.transform;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import java.util.ArrayList;
 import java.util.concurrent.Semaphore;
 
 import org.junit.Test;
-import org.riekr.jloga.io.TempTextSource;
+import org.riekr.jloga.io.VolatileTextSource;
 
 public class HeaderDetectorTest {
 
 	@Test
 	public void test() throws InterruptedException {
 		Semaphore semaphore = new Semaphore(0);
-		try (TempTextSource t = new TempTextSource()) {
-			t.addLine(0, "A|B|C|D");
-			t.addLine(0, "1|2|3|4");
-			t.addLine(0, "1|2|3|");
-			t.complete();
+		try (VolatileTextSource t = new VolatileTextSource(new ArrayList<>())) {
+			t.add("A|B|C|D");
+			t.add("1|2|3|4");
+			t.add("1|2|3|");
 			HeaderDetector detector = new HeaderDetector(t, semaphore::release, null);
 			detector.detect();
 			semaphore.acquire();